new file:   .gitignore 	modified:   test_01.py

+++ test_01.py
@@ -6,10 +6,11 @@ def rename_files(directory, str_search, str_old, str_new):
     for filename_old in os.listdir(directory):
         if re.search(rf'{str_search}', filename_old):
             filename_new = re.sub(rf'{str_old}', rf'{str_new}', filename_old)
-            print(f"Renaming {filename_old} to {filename_new}")
-        path_old = os.path.join(directory, filename_old)
-        path_new = os.path.join(directory, filename_new)
-        os.rename(path_old, path_new)
+            path_old = os.path.join(directory, filename_old)
+            path_new = os.path.join(directory, filename_new)
+            if not os.path.exists(path_new):
+                os.rename(path_old, path_new)
+                print(f"Renaming {filename_old} to {filename_new}")
 
 # Example usage:
 # rename_files("/path/to/directory", "search", "old", "new")
